# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Нууу очень крутая апишка",
        "description": "Документация для REST API Laravel",
        "contact": {
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/",
            "description": "Локальный сервер разработки"
        }
    ],
    "paths": {
...
        "/api/v1/requests": {
            "get": {
                "tags": [
                    "Requests"
                ],
                "summary": "Получить список всех заявок",
                "operationId": "e8ba1b0f112480d4aad0b1360cb1a20a",
                "responses": {
                    "200": {
                        "description": "Список заявок",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Request"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Requests"
                ],
                "summary": "Создать новую заявку",
                "operationId": "927dd34168a9440882d2d9d4aea3091e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "category",
                                    "name",
                                    "phone"
                                ],
                                "properties": {
                                    "category": {
                                        "type": "string",
                                        "enum": [
                                            "estimate",
                                            "development",
                                            "callback"
                                        ],
                                        "example": "estimate"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Александр"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "+79995553322"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Описание запроса"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Заявка успешно создана",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Request"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ошибка валидации"
                    }
                }
            }
        },
        "/api/v1/requests/{id}": {
            "get": {
                "tags": [
                    "Requests"
                ],
                "summary": "Получить заявку по ID",
                "operationId": "ccb3309007cab7ed5d86ecc7c17cb037",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID заявки",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация о заявке",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Request"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Заявка не найдена"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Requests"
                ],
                "summary": "Удалить заявку",
                "operationId": "43c0f1596e3306070bad8ecc641774f8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID заявки",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Заявка удалена"
                    },
                    "404": {
                        "description": "Заявка не найдена"
                    }
                }
            }
        },
        "/api/v1/requests/{id}/status": {
            "patch": {
                "tags": [
                    "Requests"
                ],
                "summary": "Обновить статус заявки",
                "operationId": "7aac3f26c2c6ab5634ed051413ce2264",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID заявки",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "new",
                                            "in_progress",
                                            "review",
                                            "done"
                                        ],
                                        "example": "done"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Статус обновлён"
                    },
                    "404": {
                        "description": "Заявка не найдена"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Request": {
                "title": "Request",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "category": {
                        "type": "string",
                        "enum": [
                            "estimate",
                            "development",
                            "callback"
                        ]
                    },
                    "name": {
                        "type": "string",
                        "example": "Иван"
                    },
                    "phone": {
                        "type": "string",
                        "example": "+79995553322"
                    },
                    "description": {
                        "type": "string",
                        "example": "Описание заявки"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "new",
                            "in_progress",
                            "review",
                            "done"
                        ]
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    },
...
}